requirements.txt
Django>=5.0
psycopg2-binary>=2.9.0
python-dotenv==1.0.1
django-bootstrap5
django-widget-tweaks
requests==2.32.3
google-generativeai==0.8.3
pillow==11.0.0
beautifulsoup4==4.12.3
gunicorn